requirements.txt
flask
rembg
Pillow
werkzeug
opencv-python
numpy
